+++ package.json
@@ -18,6 +18,7 @@
         "vite": "^4.0.0"
     },
     "dependencies": {
-        "@material-ui/core": "^4.12.4"
+        "@material-ui/core": "^4.12.4",
+        "react-router-dom": "^6.16.0"
     }
 }
