+++ resources/js/app.js
@@ -12,4 +12,4 @@ import './bootstrap';
  * or customize the JavaScript scaffolding to fit your unique needs.
  */
 
-import './components/Example';
+import './route';
