+++ resources/views/app.blade.php
@@ -1,6 +1,7 @@
 <!doctype html>
 <html lang="{{ str_replace('_', '-', app()->getLocale()) }}">
 <head>
+    <base href="/" >
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
