+++ routes/web.php
@@ -13,6 +13,6 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/', function () {
+Route::get('{any}', function () {
     return view('app');
-});
+})->where('any','.*');
